# Optimize series response format by using repeated fileds.