--- a/pkg/ingester/ingester_test.go
+++ b/pkg/ingester/ingester_test.go
@@ -279,16 +279,16 @@ func TestIngester(t *testing.T) {
 	require.Nil(t, err)
 	require.ElementsMatch(t, []logproto.SeriesIdentifier{
 		{
-			Labels: map[string]string{
-				"foo": "bar",
-				"bar": "baz1",
-			},
+			Labels: logproto.NewSeriesEntries(
+				"bar", "baz1",
+				"foo", "bar",
+			),
 		},
 		{
-			Labels: map[string]string{
-				"foo": "bar",
-				"bar": "baz2",
-			},
+			Labels: logproto.NewSeriesEntries(
+				"bar", "baz2",
+				"foo", "bar",
+			),
 		},
 	}, resp.GetSeries())
 
@@ -317,16 +317,16 @@ func TestIngester(t *testing.T) {
 	require.Nil(t, err)
 	require.ElementsMatch(t, []logproto.SeriesIdentifier{
 		{
-			Labels: map[string]string{
-				"foo": "bar",
-				"bar": "baz1",
-			},
+			Labels: logproto.NewSeriesEntries(
+				"bar", "baz1",
+				"foo", "bar",
+			),
 		},
 		{
-			Labels: map[string]string{
-				"foo": "bar",
-				"bar": "baz2",
-			},
+			Labels: logproto.NewSeriesEntries(
+				"bar", "baz2",
+				"foo", "bar",
+			),
 		},
 	}, resp.GetSeries())
 
@@ -339,10 +339,10 @@ func TestIngester(t *testing.T) {
 	require.Nil(t, err)
 	require.ElementsMatch(t, []logproto.SeriesIdentifier{
 		{
-			Labels: map[string]string{
-				"foo": "bar",
-				"bar": "baz2",
-			},
+			Labels: logproto.NewSeriesEntries(
+				"bar", "baz2",
+				"foo", "bar",
+			),
 		},
 	}, resp.GetSeries())
 
@@ -355,16 +355,16 @@ func TestIngester(t *testing.T) {
 	require.Nil(t, err)
 	require.ElementsMatch(t, []logproto.SeriesIdentifier{
 		{
-			Labels: map[string]string{
-				"foo": "bar",
-				"bar": "baz1",
-			},
+			Labels: logproto.NewSeriesEntries(
+				"bar", "baz1",
+				"foo", "bar",
+			),
 		},
 		{
-			Labels: map[string]string{
-				"foo": "bar",
-				"bar": "baz2",
-			},
+			Labels: logproto.NewSeriesEntries(
+				"bar", "baz2",
+				"foo", "bar",
+			),
 		},
 	}, resp.GetSeries())
 }