--- a/pkg/ingester/instance.go
+++ b/pkg/ingester/instance.go
@@ -572,9 +572,11 @@ func (i *instance) Series(ctx context.Context, req *logproto.SeriesRequest) (*lo
 		err = i.forMatchingStreams(ctx, req.Start, nil, shard, func(stream *stream) error {
 			// consider the stream only if it overlaps the request time range
 			if shouldConsiderStream(stream, req.Start, req.End) {
-				series = append(series, logproto.SeriesIdentifier{
-					Labels: stream.labels.Map(),
-				})
+				id := logproto.SeriesIdentifier{}
+				for _, l := range stream.labels {
+					id.Labels = append(id.Labels, &logproto.SeriesIdentifier_LabelsEntry{Key: l.Name, Value: l.Value})
+				}
+				series = append(series, id)
 			}
 			return nil
 		})
@@ -597,9 +599,11 @@ func (i *instance) Series(ctx context.Context, req *logproto.SeriesRequest) (*lo
 						return nil
 					}
 
-					dedupedSeries[key] = logproto.SeriesIdentifier{
-						Labels: stream.labels.Map(),
+					id := logproto.SeriesIdentifier{}
+					for _, l := range stream.labels {
+						id.Labels = append(id.Labels, &logproto.SeriesIdentifier_LabelsEntry{Key: l.Name, Value: l.Value})
 					}
+					dedupedSeries[key] = id
 				}
 				return nil
 			})