--- a/pkg/ingester/instance.go
+++ b/pkg/ingester/instance.go
@@ -572,9 +572,11 @@ func (i *instance) Series(ctx context.Context, req *logproto.SeriesRequest) (*lo
 		err = i.forMatchingStreams(ctx, req.Start, nil, shard, func(stream *stream) error {
 			// consider the stream only if it overlaps the request time range
 			if shouldConsiderStream(stream, req.Start, req.End) {
-				id := logproto.SeriesIdentifier{}
-				for _, l := range stream.labels {
-					id.Labels = append(id.Labels, &logproto.SeriesIdentifier_LabelsEntry{Key: l.Name, Value: l.Value})
+				id := logproto.SeriesIdentifier{
+					Labels: make([]logproto.SeriesIdentifier_LabelsEntry, len(stream.labels)),
+				}
+				for i, l := range stream.labels {
+					id.Labels[i] = logproto.SeriesIdentifier_LabelsEntry{Key: l.Name, Value: l.Value}
 				}
 				series = append(series, id)
 			}
@@ -599,9 +601,11 @@ func (i *instance) Series(ctx context.Context, req *logproto.SeriesRequest) (*lo
 						return nil
 					}
 
-					id := logproto.SeriesIdentifier{}
-					for _, l := range stream.labels {
-						id.Labels = append(id.Labels, &logproto.SeriesIdentifier_LabelsEntry{Key: l.Name, Value: l.Value})
+					id := logproto.SeriesIdentifier{
+						Labels: make([]logproto.SeriesIdentifier_LabelsEntry, len(stream.labels)),
+					}
+					for i, l := range stream.labels {
+						id.Labels[i] = logproto.SeriesIdentifier_LabelsEntry{Key: l.Name, Value: l.Value}
 					}
 					dedupedSeries[key] = id
 				}