--- a/pkg/querier/multi_tenant_querier.go
+++ b/pkg/querier/multi_tenant_querier.go
@@ -160,11 +160,10 @@ func (q *MultiTenantQuerier) Series(ctx context.Context, req *logproto.SeriesReq
 			return nil, err
 		}
 
-		for _, s := range resp.GetSeries() {
-			for _, pair := range s.Labels {
-				if pair.Key == defaultTenantLabel {
-					pair.Value = id
-				}
+		for i := range resp.GetSeries() {
+			s := &resp.Series[i]
+			if s.Get(defaultTenantLabel) == "" {
+				s.Labels = append(s.Labels, &logproto.SeriesIdentifier_LabelsEntry{Key: defaultTenantLabel, Value: id})
 			}
 		}