--- a/pkg/querier/queryrange/codec.go
+++ b/pkg/querier/queryrange/codec.go
@@ -866,10 +866,10 @@ func decodeResponseJSONFrom(buf []byte, req queryrangebase.Request, headers http
 		data := make([]logproto.SeriesIdentifier, 0, len(resp.Data))
 		for _, label := range resp.Data {
 			d := logproto.SeriesIdentifier{
-				Labels: make([]*logproto.SeriesIdentifier_LabelsEntry, 0, len(label)),
+				Labels: make([]logproto.SeriesIdentifier_LabelsEntry, 0, len(label)),
 			}
 			for n, v := range label {
-				d.Labels = append(d.Labels, &logproto.SeriesIdentifier_LabelsEntry{Key: n, Value: v})
+				d.Labels = append(d.Labels, logproto.SeriesIdentifier_LabelsEntry{Key: n, Value: v})
 			}
 			data = append(data, d)
 		}