--- a/pkg/querier/queryrange/codec.go
+++ b/pkg/querier/queryrange/codec.go
@@ -858,28 +858,15 @@ func decodeResponseJSONFrom(buf []byte, req queryrangebase.Request, headers http
 
 	switch req := req.(type) {
 	case *LokiSeriesRequest:
-		var resp loghttp.SeriesResponse
+		resp := &LokiSeriesResponse{
+			Version: uint32(loghttp.GetVersion(req.Path)),
+			Headers: httpResponseHeadersToPromResponseHeaders(headers),
+		}
 		if err := json.Unmarshal(buf, &resp); err != nil {
 			return nil, httpgrpc.Errorf(http.StatusInternalServerError, "error decoding response: %v", err)
 		}
 
-		data := make([]logproto.SeriesIdentifier, 0, len(resp.Data))
-		for _, label := range resp.Data {
-			d := logproto.SeriesIdentifier{
-				Labels: make([]logproto.SeriesIdentifier_LabelsEntry, 0, len(label)),
-			}
-			for n, v := range label {
-				d.Labels = append(d.Labels, logproto.SeriesIdentifier_LabelsEntry{Key: n, Value: v})
-			}
-			data = append(data, d)
-		}
-
-		return &LokiSeriesResponse{
-			Status:  resp.Status,
-			Version: uint32(loghttp.GetVersion(req.Path)),
-			Data:    data,
-			Headers: httpResponseHeadersToPromResponseHeaders(headers),
-		}, nil
+		return resp, nil
 	case *LabelRequest:
 		var resp loghttp.LabelResponse
 		if err := json.Unmarshal(buf, &resp); err != nil {