--- a/pkg/querier/queryrange/views.go
+++ b/pkg/querier/queryrange/views.go
@@ -304,7 +304,7 @@ func (v *MergedSeriesResponseView) Materialize() (*LokiSeriesResponse, error) {
 	err := v.ForEachUniqueSeries(func(series *SeriesIdentifierView) error {
 		identifier := logproto.SeriesIdentifier{Labels: make([]*logproto.SeriesIdentifier_LabelsEntry, 0)}
 		err := series.ForEachLabel(func(name, value string) error {
-			// TODO: identifier.Labels[name] = value
+			identifier.Labels = append(identifier.Labels, &logproto.SeriesIdentifier_LabelsEntry{Key: name, Value: value})
 			return nil
 		})
 		if err != nil {