--- a/pkg/querier/queryrange/views.go
+++ b/pkg/querier/queryrange/views.go
@@ -302,9 +302,9 @@ func (v *MergedSeriesResponseView) ForEachUniqueSeries(fn func(*SeriesIdentifier
 func (v *MergedSeriesResponseView) Materialize() (*LokiSeriesResponse, error) {
 	mat := &LokiSeriesResponse{}
 	err := v.ForEachUniqueSeries(func(series *SeriesIdentifierView) error {
-		identifier := logproto.SeriesIdentifier{Labels: make([]*logproto.SeriesIdentifier_LabelsEntry, 0)}
+		identifier := logproto.SeriesIdentifier{Labels: make([]logproto.SeriesIdentifier_LabelsEntry, 0)}
 		err := series.ForEachLabel(func(name, value string) error {
-			identifier.Labels = append(identifier.Labels, &logproto.SeriesIdentifier_LabelsEntry{Key: name, Value: value})
+			identifier.Labels = append(identifier.Labels, logproto.SeriesIdentifier_LabelsEntry{Key: name, Value: value})
 			return nil
 		})
 		if err != nil {