--- a/pkg/storage/store.go
+++ b/pkg/storage/store.go
@@ -454,9 +454,11 @@ func (s *LokiStore) SelectSeries(ctx context.Context, req logql.SelectLogParams)
 	}
 	result := make([]logproto.SeriesIdentifier, len(series))
 	for i, s := range series {
-		result[i] = logproto.SeriesIdentifier{}
-		for _, l := range s {
-			result[i].Labels = append(result[i].Labels, &logproto.SeriesIdentifier_LabelsEntry{Key: l.Name, Value: l.Value})
+		result[i] = logproto.SeriesIdentifier{
+			Labels: make([]logproto.SeriesIdentifier_LabelsEntry, len(s)),
+		}
+		for j, l := range s {
+			result[i].Labels[j] = logproto.SeriesIdentifier_LabelsEntry{Key: l.Name, Value: l.Value}
 		}
 	}
 	return result, nil