**What this PR does / why we need it**:
The Protobuf map type is encodied as a repeated field of map entries. Decoding them to a slice is much faster than decoding them into a map. Since Loki is not using the fast key check for a map we can use the slice decoding.

This change also allows us to decode the JSON directly into the right protobuf struct. This doulbes the JSON decoding speed and reduces the memory pressure by ~40%.

```
› go test -bench=. -run=^$ -count=10 ./pkg/querier/queryrange > before.log
› go test -bench=. -run=^$ -count=10 ./pkg/querier/queryrange > after.log
› benchstat before.log after.log
goos: darwin
goarch: arm64
pkg: github.com/grafana/loki/pkg/querier/queryrange
                                                             │  before.log  │              after.log               │
                                                             │    sec/op    │    sec/op     vs base                │
ResponseMerge/mergeStreams_unlimited-10                         32.36m ± 0%   32.63m ±  2%        ~ (p=0.393 n=10)
ResponseMerge/mergeOrderedNonOverlappingStreams_unlimited-10    1.050m ± 1%   1.080m ±  3%   +2.84% (p=0.005 n=10)
ResponseMerge/mergeStreams_limited-10                           33.02m ± 0%   32.60m ±  1%   -1.29% (p=0.004 n=10)
ResponseMerge/mergeOrderedNonOverlappingStreams_limited-10      15.11m ± 0%   15.07m ±  0%        ~ (p=0.075 n=10)
_CodecDecodeLogs-10                                             4.395m ± 1%   4.364m ±  0%   -0.72% (p=0.005 n=10)
_CodecDecodeSamples-10                                          16.97m ± 0%   16.84m ±  2%   -0.77% (p=0.023 n=10)
_CodecDecodeSeries/application/vnd.google.protobuf-10           745.8µ ± 8%   736.8µ ± 12%        ~ (p=0.739 n=10)
_CodecDecodeSeries/application/json;_charset=utf-8-10           15.37m ± 1%   10.60m ±  0%  -31.03% (p=0.000 n=10)
_MergeResponses-10                                             1186.9m ± 2%   149.8m ±  1%  -87.38% (p=0.000 n=10)
_UnwrapSeries-10                                                9.399m ± 1%   4.049m ±  0%  -56.92% (p=0.000 n=10)
_DecodeMergeEncodeCycle-10                                      666.0m ± 3%   194.6m ±  0%  -70.79% (p=0.000 n=10)
geomean                                                         18.87m        12.51m        -33.70%

                                                      │  before.log   │              after.log               │
                                                      │     B/op      │     B/op      vs base                │
_CodecDecodeLogs-10                                      3.649Mi ± 0%   3.649Mi ± 0%        ~ (p=0.364 n=10)
_CodecDecodeSamples-10                                   18.12Mi ± 0%   18.12Mi ± 0%        ~ (p=0.926 n=10)
_CodecDecodeSeries/application/vnd.google.protobuf-10    7.647Mi ± 0%   7.647Mi ± 0%        ~ (p=0.587 n=10)
_CodecDecodeSeries/application/json;_charset=utf-8-10    27.94Mi ± 0%   16.99Mi ± 0%  -39.18% (p=0.000 n=10)
_MergeResponses-10                                       2.362Mi ± 0%   2.408Mi ± 0%   +1.98% (p=0.000 n=10)
_UnwrapSeries-10                                        19.495Mi ± 0%   8.595Mi ± 0%  -55.91% (p=0.000 n=10)
_DecodeMergeEncodeCycle-10                               772.3Mi ± 0%   772.3Mi ± 0%        ~ (p=0.912 n=10)
geomean                                                  17.50Mi        14.54Mi       -16.91%

                                                      │ before.log  │              after.log               │
                                                      │  allocs/op  │  allocs/op   vs base                 │
_CodecDecodeLogs-10                                     41.10k ± 0%   41.10k ± 0%       ~ (p=1.000 n=10) ¹
_CodecDecodeSamples-10                                  411.9k ± 0%   411.9k ± 0%       ~ (p=1.000 n=10)
_CodecDecodeSeries/application/vnd.google.protobuf-10    32.00 ± 0%    32.00 ± 0%       ~ (p=1.000 n=10) ¹
_CodecDecodeSeries/application/json;_charset=utf-8-10   304.2k ± 0%   298.1k ± 0%  -2.01% (p=0.000 n=10)
_MergeResponses-10                                      100.1k ± 0%   100.1k ± 0%  -0.00% (p=0.002 n=10)
_UnwrapSeries-10                                        201.1k ± 0%   198.0k ± 0%  -1.54% (p=0.000 n=10)
_DecodeMergeEncodeCycle-10                              203.1k ± 0%   203.1k ± 0%       ~ (p=0.621 n=10)
geomean                                                 48.95k        48.70k       -0.51%
¹ all samples are equal
```

**Checklist**
- [x] Reviewed the [`CONTRIBUTING.md`](https://github.com/grafana/loki/blob/main/CONTRIBUTING.md) guide (**required**)
- [ ] Documentation added
- [x] Tests updated
- [ ] `CHANGELOG.md` updated
  - [ ] If the change is worth mentioning in the release notes, add `add-to-release-notes` label
- [ ] Changes that require user attention or interaction to upgrade are documented in `docs/sources/setup/upgrade/_index.md`
- [ ] For Helm chart changes bump the Helm chart version in `production/helm/loki/Chart.yaml` and update `production/helm/loki/CHANGELOG.md` and `production/helm/loki/README.md`. [Example PR](https://github.com/grafana/loki/commit/d10549e3ece02120974929894ee333d07755d213)
- [ ] If the change is deprecating or removing a configuration option, update the `deprecated-config.yaml` and `deleted-config.yaml` files respectively in the `tools/deprecated-config-checker` directory. [Example PR](https://github.com/grafana/loki/pull/10840/commits/0d4416a4b03739583349934b96f272fb4f685d15)