@jeschkies 2023-12-18
> MergeResponses-10 needs some investigation!

Turns out that `sort.SliceFunc` is using temporary memory fro swapping elements.