@jeschkies 2023-12-18
> I'm surprised by the benchmark numbers, tbh.

@chaudum what is surprising to you?