@chaudum 2023-12-18
> > I'm surprised by the benchmark numbers, tbh.
> 
> @chaudum what is surprising to you?

> geomean                                                         18.87m        12.51m        -33.70%