@jeschkies 2023-12-19
> geomean 18.87m 12.51m -33.70%

That's because it's faster to iterate through a slice and the allocation is faster. Sometimes we just need to allocate one slice. Btw we save a lot of allocations by changing the slice into a slice of struct instead of a slice of pointers.